Add files via upload
Mobile Price predition using machine learning algorithms 

Algorithms Used : Logistic Regression, Decision Tree, Random Forest 
Dataset: Mobile Price prediction (Source: Kaggle) 

The dataset has details about the various attributes that influences the price range of the mobiles. 

In this dataset we have 2000 rows and 21 columns to inspect, since the attributes are binary classified, even though we need to dig in various attributes that influences the 
price range 

Here in this dataset, we have battery power and ram columns which influences the price alot, so we need to proceed with EDA and we need to encode the battery power attribute 
to understand how the multiple price factors are evenly distributed in this dataset.